Moves files into packages, removes route object, and removes unecessary items for router
