+++ src/main/java/uk/nickbdyer/httpserver/requests/Method.java
@@ -1,4 +1,4 @@
-package uk.nickbdyer.httpserver;
+package uk.nickbdyer.httpserver.requests;
 
 public enum Method {
     POST, HEAD, INVALID_METHOD, PUT, DELETE, TRACE, CONNECT, OPTIONS, GET
