+++ src/main/java/uk/nickbdyer/httpserver/requests/RequestLine.java
@@ -1,4 +1,4 @@
-package uk.nickbdyer.httpserver;
+package uk.nickbdyer.httpserver.requests;
 
 public class RequestLine {
