+++ src/test/java/uk/nickbdyer/httpserver/ResponseDispatcherTest.java
@@ -1,6 +1,7 @@
 package uk.nickbdyer.httpserver;
 
 import org.junit.Test;
+import uk.nickbdyer.httpserver.responses.ResponseDispatcher;
 import uk.nickbdyer.httpserver.testdoubles.SocketStubWithOutputStream;
 
 import java.io.ByteArrayOutputStream;
