+++ src/test/java/uk/nickbdyer/httpserver/controllers/RedirectControllerTest.java
@@ -1,11 +1,11 @@
 package uk.nickbdyer.httpserver.controllers;
 
 import org.junit.Test;
-import uk.nickbdyer.httpserver.Request;
-import uk.nickbdyer.httpserver.Response;
+import uk.nickbdyer.httpserver.requests.Request;
+import uk.nickbdyer.httpserver.responses.Response;
 
 import static org.junit.Assert.assertEquals;
-import static uk.nickbdyer.httpserver.Method.GET;
+import static uk.nickbdyer.httpserver.requests.Method.GET;
 
 public class RedirectControllerTest {
